requirements.txt
boto3==1.9.138
botocore==1.12.138
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cognitojwt==1.1.0
docutils==0.14
ecdsa==0.13.2
envs==1.3
Flask==1.0.2
Flask-Cognito==1.13
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
pyasn1==0.4.5
pycryptodome==3.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-jose==3.0.1
python-jose-cryptodome==1.3.2
requests==2.21.0
rsa==4.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.2
warrant==0.6.1
Werkzeug==0.15.2
